build: produce CommonJS bundle instead of UMD
UMD seems to make sense only if we use it in browser without any
bundler, but in order to work, it requires all dependencies to be either
inlined or to be attached to globals.

In the case of "widget-utils", we depend on "lodash", "moment-range" and
"moment-timezone" and neither inline them to the "widget-utils" bundle
nor access them using the global scope.

We use `require()` to consume these dependencies, and that's a clear
indication that we require "widget-utils" to be executed in CommonJS
environment anyway. So there's no need to generate UMD in such a case.